pyproject.toml
[tool.poetry]
name = "runner-bot"
version = "0.1.0"
description = ""
authors = ["Arash Outadi <arash.out@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.6.0"
pyowm = "^3.1.1"
dateparser = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
